+++ Editor/AddressableImportRule.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d5de26755a197ab439bbb3a6ac4809c8
+guid: 54d606711a1ab2e45be6335860cfb3f4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
