+++ Editor/AddressableImportSettings.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4457b6e6d3304bd46a31be63dad3325c
+guid: e261d7bf09cd6f143bf70187609b1d80
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
