+++ Editor/AddressableImporter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a88f0d7d780c7824daec928bfe13da4f
+guid: 83025f04d5722a9419b458539c30cc8b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
