+++ Editor/Helper/AddressablesImporterOdinInspectorDrawer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4c668f9525a32f34aa408a6384243a0a
+guid: 36b62799a4bc2d14ea554520750092db
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
