+++ Editor/Unity.AddressableImporter.Editor.asmdef.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6256bdcb0aaf53446b7dd10403a9d748
+guid: 636b60a8741064d4b96265e97d4002a7
 AssemblyDefinitionImporter:
   externalObjects: {}
   userData:
