+++ Media/icon-128.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3c8cb60432c8c0b4bb10cc05ca10b86c
+guid: 78827eb86b20b4a40ba74b082a59dbd9
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
