+++ Media/icon-512.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 67a1a903f0792a84295bd9892f6e0a6f
+guid: e2c244545a4cff9498de906a17105477
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
