+++ Media/icon.svg.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f7307b5665f31414a919ea167c5a59e6
+guid: e6abc220e29f8644ab36a67184b9e1b1
 DefaultImporter:
   externalObjects: {}
   userData:
