+++ Media/ogimage.psd.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 13e57e4c581efe4439cda4c7a8969a3e
+guid: 8f64debee5c74024391d4558062424ba
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
