+++ Runtime/ButtonMethodAttribute.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+
+namespace UnityAddressableImporter.Helper
+{
+    using System;
+
+    [AttributeUsage(AttributeTargets.Method)]
+    public class ButtonMethodAttribute : PropertyAttribute
+    {
+    }
+}
+
