+++ Tests/Editor/AddressableImportRuleTests.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0805d3b7bcf8e88488938813fd44d6ef
+guid: 4271cb5dfcd584e41a443daff083228d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
