+++ Tests/Editor/Unity.AddressableImporter.Editor.Tests.asmdef.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1ddb623ed29d1984d93977489a666407
+guid: de73af9ef85ad1c44b90dd16d7d77899
 AssemblyDefinitionImporter:
   externalObjects: {}
   userData:
